backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
firebase-admin==6.5.0
pyjwt[crypto]==2.8.0
google-api-python-client==2.129.0
google-auth-oauthlib==1.2.0
google-cloud-storage==2.16.0
anthropic==0.26.0
reportlab==4.2.0
pytest==8.2.1
pytest-asyncio==0.23.6
httpx==0.27.0
